api/pyproject.toml
[tool.poetry]
package-mode = false
name = "data-sharing-barter-incentives-rest-api"
version = "0.1.0"
description = "Restful API and Database for INESC TEC Data Sharing / Barter Incentives Service"
authors = ["Jose Andrade <jose.r.andrade@inesctec.pt>", "Andre Garcia <andre.f.garcia@inesctec.pt>"]
license = "GNU Affero General Public License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.3"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
PyJWT = "^2.8.0"
pytz = "^2024.1"
psycopg2-binary = "^2.9.6"
pandas = "^2.2.1"
loguru = "^0.7.2"
gunicorn = "^21.2.0"
django-stronghold = "^0.4.0"
drf-yasg = "^1.21.7"
django-structlog = "^7.1.0"
django-cors-headers = "^4.3.1"
python-dotenv = "^1.0.1"
iota-sdk = "^1.1.3"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-django = "^4.8.0"
pytest-cov = "^4.1.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
